client: Refactor message serialization tests
- Refactor the tests to use generic test functions.

- In preparation for extending these tests to newer wire serialization
  adapters.
