+++ client/proposalmsgs_test.go
@@ -1,4 +1,4 @@
-// Copyright 2019 - See NOTICE file for copyright holders.
+// Copyright 2022 - See NOTICE file for copyright holders.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -15,7 +15,6 @@
 package client_test
 
 import (
-	"math/rand"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -25,11 +24,13 @@ import (
 	"perun.network/go-perun/channel/test"
 	"perun.network/go-perun/client"
 	clienttest "perun.network/go-perun/client/test"
-	wallettest "perun.network/go-perun/wallet/test"
-	"perun.network/go-perun/wire"
 	peruniotest "perun.network/go-perun/wire/perunio/test"
 	pkgtest "polycry.pt/poly-go/test"
 )
+
+func TestProposalMsgsSerialization(t *testing.T) {
+	clienttest.ProposalMsgsSerializationTest(t, peruniotest.MsgSerializerTest)
+}
 
 func TestNewLedgerChannelProposal(t *testing.T) {
 	rng := pkgtest.Prng(t)
@@ -49,31 +50,6 @@ func TestNewLedgerChannelProposal(t *testing.T) {
 	agreement = test.NewRandomBalances(rng, test.WithNumAssets(len(base.InitBals.Assets)))
 	_, err = client.NewLedgerChannelProposal(base.ChallengeDuration, base.Participant, base.InitBals, base.Peers, client.WithFundingAgreement(agreement))
 	assert.EqualError(t, err, "FundingAgreement and initial balances differ")
-}
-
-func TestChannelProposalReqSerialization(t *testing.T) {
-	rng := pkgtest.Prng(t)
-	for i := 0; i < 16; i++ {
-		var (
-			app client.ProposalOpts
-			m   wire.Msg
-			err error
-		)
-		if i&1 == 0 {
-			app = client.WithApp(test.NewRandomAppAndData(rng))
-		}
-		switch i % 3 {
-		case 0:
-			m = clienttest.NewRandomLedgerChannelProposal(rng, client.WithNonceFrom(rng), app)
-		case 1:
-			m, err = clienttest.NewRandomSubChannelProposal(rng, client.WithNonceFrom(rng), app)
-			require.NoError(t, err)
-		case 2:
-			m, err = clienttest.NewRandomVirtualChannelProposal(rng, client.WithNonceFrom(rng), app)
-			require.NoError(t, err)
-		}
-		peruniotest.MsgSerializerTest(t, m)
-	}
 }
 
 func TestLedgerChannelProposalReqProposalID(t *testing.T) {
@@ -110,69 +86,4 @@ func TestVirtualChannelProposalReqProposalID(t *testing.T) {
 	assert.Equal(t, channel.NoData(), fake.InitData, "virtual channel should always has no data")
 
 	assert.NotEqual(t, original.ProposalID, fake.ProposalID)
-}
-
-func TestChannelProposalAccSerialization(t *testing.T) {
-	rng := pkgtest.Prng(t)
-	t.Run("ledger channel", func(t *testing.T) {
-		for i := 0; i < 16; i++ {
-			proposal := clienttest.NewRandomLedgerChannelProposal(rng)
-			m := proposal.Accept(
-				wallettest.NewRandomAddress(rng),
-				client.WithNonceFrom(rng))
-			peruniotest.MsgSerializerTest(t, m)
-		}
-	})
-	t.Run("sub channel", func(t *testing.T) {
-		for i := 0; i < 16; i++ {
-			var err error
-			proposal, err := clienttest.NewRandomSubChannelProposal(rng)
-			require.NoError(t, err)
-			m := proposal.Accept(client.WithNonceFrom(rng))
-			peruniotest.MsgSerializerTest(t, m)
-		}
-	})
-	t.Run("virtual channel", func(t *testing.T) {
-		for i := 0; i < 16; i++ {
-			var err error
-			proposal, err := clienttest.NewRandomVirtualChannelProposal(rng)
-			require.NoError(t, err)
-			m := proposal.Accept(wallettest.NewRandomAddress(rng))
-			peruniotest.MsgSerializerTest(t, m)
-		}
-	})
-}
-
-func TestChannelProposalRejSerialization(t *testing.T) {
-	rng := pkgtest.Prng(t)
-	for i := 0; i < 16; i++ {
-		m := &client.ChannelProposalRejMsg{
-			ProposalID: newRandomProposalID(rng),
-			Reason:     newRandomString(rng, 16, 16),
-		}
-		peruniotest.MsgSerializerTest(t, m)
-	}
-}
-
-func TestSubChannelProposalSerialization(t *testing.T) {
-	rng := pkgtest.Prng(t)
-	const repeatRandomizedTest = 16
-	for i := 0; i < repeatRandomizedTest; i++ {
-		prop, err := clienttest.NewRandomSubChannelProposal(rng)
-		require.NoError(t, err)
-		peruniotest.MsgSerializerTest(t, prop)
-	}
-}
-
-func newRandomProposalID(rng *rand.Rand) (id client.ProposalID) {
-	rng.Read(id[:])
-	return
-}
-
-// newRandomstring returns a random string of length between minLen and
-// minLen+maxLenDiff.
-func newRandomString(rng *rand.Rand, minLen, maxLenDiff int) string {
-	r := make([]byte, minLen+rng.Intn(maxLenDiff))
-	rng.Read(r)
-	return string(r)
 }
