+++ client/updatemsgs_test.go
@@ -1,4 +1,4 @@
-// Copyright 2019 - See NOTICE file for copyright holders.
+// Copyright 2022 - See NOTICE file for copyright holders.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -15,117 +15,12 @@
 package client_test
 
 import (
-	"math/rand"
 	"testing"
 
-	"perun.network/go-perun/channel"
-	"perun.network/go-perun/channel/test"
-	"perun.network/go-perun/client"
-	"perun.network/go-perun/wallet"
-	wallettest "perun.network/go-perun/wallet/test"
+	clienttest "perun.network/go-perun/client/test"
 	peruniotest "perun.network/go-perun/wire/perunio/test"
-	pkgtest "polycry.pt/poly-go/test"
 )
 
-func TestChannelUpdateSerialization(t *testing.T) {
-	rng := pkgtest.Prng(t)
-	for i := 0; i < 4; i++ {
-		m := newRandomMsgChannelUpdate(rng)
-		peruniotest.MsgSerializerTest(t, m)
-	}
-}
-
-func newRandomMsgChannelUpdate(rng *rand.Rand) *client.ChannelUpdateMsg {
-	state := test.NewRandomState(rng)
-	sig := newRandomSig(rng)
-	return &client.ChannelUpdateMsg{
-		ChannelUpdate: client.ChannelUpdate{
-			State:    state,
-			ActorIdx: channel.Index(rng.Intn(state.NumParts())),
-		},
-		Sig: sig,
-	}
-}
-
-func TestSerialization_VirtualChannelFundingProposal(t *testing.T) {
-	rng := pkgtest.Prng(t)
-	for i := 0; i < 4; i++ {
-		msgUp := newRandomMsgChannelUpdate(rng)
-		params, state := test.NewRandomParamsAndState(rng)
-		m := &client.VirtualChannelFundingProposalMsg{
-			ChannelUpdateMsg: *msgUp,
-			Initial: channel.SignedState{
-				Params: params,
-				State:  state,
-				Sigs:   newRandomSigs(rng, state.NumParts()),
-			},
-			IndexMap: test.NewRandomIndexMap(rng, state.NumParts(), msgUp.State.NumParts()),
-		}
-		peruniotest.MsgSerializerTest(t, m)
-	}
-}
-
-func TestSerialization_VirtualChannelSettlementProposal(t *testing.T) {
-	rng := pkgtest.Prng(t)
-	for i := 0; i < 4; i++ {
-		msgUp := newRandomMsgChannelUpdate(rng)
-		params, state := test.NewRandomParamsAndState(rng)
-		m := &client.VirtualChannelSettlementProposalMsg{
-			ChannelUpdateMsg: *msgUp,
-			Final: channel.SignedState{
-				Params: params,
-				State:  state,
-				Sigs:   newRandomSigs(rng, state.NumParts()),
-			},
-		}
-		peruniotest.MsgSerializerTest(t, m)
-	}
-}
-
-func TestChannelUpdateAccSerialization(t *testing.T) {
-	rng := pkgtest.Prng(t)
-	for i := 0; i < 4; i++ {
-		sig := newRandomSig(rng)
-		m := &client.ChannelUpdateAccMsg{
-			ChannelID: test.NewRandomChannelID(rng),
-			Version:   uint64(rng.Int63()),
-			Sig:       sig,
-		}
-		peruniotest.MsgSerializerTest(t, m)
-	}
-}
-
-func TestChannelUpdateRejSerialization(t *testing.T) {
-	rng := pkgtest.Prng(t)
-	for i := 0; i < 4; i++ {
-		m := &client.ChannelUpdateRejMsg{
-			ChannelID: test.NewRandomChannelID(rng),
-			Version:   uint64(rng.Int63()),
-			Reason:    "random-string",
-			// Reason:    newRandomString(rng, 16, 16),
-		}
-		peruniotest.MsgSerializerTest(t, m)
-	}
-}
-
-// newRandomSig generates a random account and then returns the signature on
-// some random data.
-func newRandomSig(rng *rand.Rand) wallet.Sig {
-	acc := wallettest.NewRandomAccount(rng)
-	data := make([]byte, 8)
-	rng.Read(data)
-	sig, err := acc.SignData(data)
-	if err != nil {
-		panic("signing error")
-	}
-	return sig
-}
-
-// newRandomSigs generates a list of random signatures.
-func newRandomSigs(rng *rand.Rand, n int) (a []wallet.Sig) {
-	a = make([]wallet.Sig, n)
-	for i := range a {
-		a[i] = newRandomSig(rng)
-	}
-	return
+func Test_UpdateMsgsSerialization(t *testing.T) {
+	clienttest.UpdateMsgsSerializationTest(t, peruniotest.MsgSerializerTest)
 }
